requirements.txt
crewai
langchain-groq
python-dotenv
exa-py
langchain_google_genai
opencv-python
